==> notebooks/01-load-and-validate-raw-data.ipynb <==
from pathlib import Path
import requests

def download_one_file_of_raw_data(year: int, month:int) -> Path:
    """Fetch Raw data from  url
    """
    url = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2022-01.parquet"
    response = requests.get(url)

    if response.status_code == 200:
        path = f"../data/raw/rides_{year}--{month:02d}.parquet"
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception(f"{url} is not available")



download_one_file_of_raw_data(year=2022, month=1)

import pandas as pd 
rides = pd.read_parquet("../data/raw/rides_2022--01.parquet")

rides.head(5)

rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides.head()

rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id',
}, inplace=True)

rides.head(10)



rides['pickup_datetime'].describe()





rides = rides[rides.pickup_datetime >= '2022-01-01']
rides = rides[rides.pickup_datetime < '2022-02-01']
rides['pickup_datetime'].describe()



rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
