==> notebooks/17_monitoring_demo.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
sys.path.append("/home/mcamara/taxi_demand_predictor")

import src.config as config

from datetime import datetime

import pandas as pd

current_date = pd.to_datetime(datetime.utcnow()).floor('H')

from src.feature_store_api import get_feature_group

predictions_fg = get_feature_group(name=config.FEATURE_GROUP_MODEL_PREDICTIONS)

actuals_fg = get_feature_group(name=config.FEATURE_GROUP_NAME)

from datetime import timedelta

query = predictions_fg.select_all().join(actuals_fg.select_all(), 
                                         on=['pickup_hour', 'pickup_location_id']).filter(predictions_fg.pickup_hour >= current_date - timedelta(days=30))

from src.feature_store_api import get_feature_store

feature_store = get_feature_store()

try:
    # create feature view as it does not exist yet\n",
    feature_store.create_feature_view( 
    name=config.FEATURE_VIEW_MONITORING, 
    version=1, 
    query=query 
    ) 
except: 
    print('Feature view already existed. Skip creation.') 
     

predictions_and_actuals_fv = feature_store.get_feature_view(name=config.FEATURE_VIEW_MONITORING, 
                                                            version=1) 

preds_and_actuals = predictions_and_actuals_fv.get_batch_data(start_time=(current_date - timedelta(days=30)),
                                                              end_time=current_date)
   

preds_and_actuals = preds_and_actuals[preds_and_actuals.pickup_hour.between(current_date - timedelta(days=30),
                                                                            current_date)]

df = preds_and_actuals

from sklearn.metrics import mean_absolute_error

mae = mean_absolute_error(df['rides'], df['predicted_demand'])
print(f'{mae=:.2f}')
